~ lookup-adjacent-slide-positions
I have determined from looking at the original javascript code I wrote
that both lookup-adjacent-slide-positions and
lookup-adjacent-climb-positions had special forms they used from within
higher-order path-searching algorithms; in both cases, it was to allow
the “starting” position (the position of the piece from which the path
originates) to be treated as though the piece had already been removed
from the board.

However, that seems silly now because with Clojure it is trivial to
simply pass in a derivative board to the beginning of these algorithms;
a board in which the piece in question no longer exists, and use its
starting position; that way, the extra parameter and logic flow can
basically be ignored, simplifying both of these functions somewhat.